feat(users):implemented DBRepository, deleted in memory repository and updated UserService

+++ backend/src/main/java/backend/repository/DBUserRepository.java
@@ -3,5 +3,5 @@ package backend.repository;
 import backend.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DBUserRepository extends JpaRepository<User, Integer> {
+public interface DBUserRepository extends JpaRepository<User, Long> {
 }
